refactor: shuffle functions
