+++ packages/app-builder/src/components/Scenario/AstBuilder/AstBuilderNode/Operand/OperandEditor/OperandEditModal/AggregationEdit/AggregationEdit.tsx
@@ -1,12 +1,12 @@
 import { Callout } from '@app-builder/components/Callout';
 import { ExternalLink } from '@app-builder/components/ExternalLink';
 import { EvaluationErrors } from '@app-builder/components/Scenario/ScenarioValidationError';
+import { type AstNode } from '@app-builder/models';
 import {
   type AggregationAstNode,
   aggregationAstNodeName,
-  type AstNode,
-  NewConstantAstNode,
-} from '@app-builder/models';
+} from '@app-builder/models/astNode/aggregation';
+import { NewConstantAstNode } from '@app-builder/models/astNode/constant';
 import {
   type AggregatorOperator,
   aggregatorOperators,
