+++ packages/app-builder/src/models/astNode/constant.ts
@@ -1,0 +1,25 @@
+import { type AstNode, type ConstantType } from './ast-node';
+
+export interface ConstantAstNode<T extends ConstantType = ConstantType> {
+  name: null;
+  constant: T;
+  children: [];
+  namedChildren: Record<string, never>;
+}
+
+export function NewConstantAstNode<T extends ConstantType = ConstantType>({
+  constant,
+}: {
+  constant: T;
+}): ConstantAstNode<T> {
+  return {
+    name: null,
+    constant: constant,
+    children: [],
+    namedChildren: {},
+  };
+}
+
+export function isConstant(node: AstNode): node is ConstantAstNode {
+  return !node.name && node.constant !== undefined;
+}
