+++ packages/app-builder/src/models/index.ts
@@ -1,4 +1,4 @@
-export * from './ast-node';
+export * from './astNode/ast-node';
 export * from './auth-errors';
 export * from './data-model';
 export * from './http-errors';
