+++ packages/app-builder/src/services/ast-node/getCustomListAccessCustomList.ts
@@ -1,4 +1,4 @@
-import { type CustomListAccessAstNode } from '@app-builder/models';
+import { type CustomListAccessAstNode } from '@app-builder/models/astNode/custom-list';
 import { type CustomList } from '@app-builder/models/custom-list';
 import * as R from 'remeda';
