+++ packages/app-builder/src/services/editor/getEnumOptionsFromNeighbour.ts
@@ -2,9 +2,9 @@ import {
   type AstNode,
   type DataModel,
   type EnumValue,
-  isDataAccessorAstNode,
   type TableModel,
 } from '@app-builder/models';
+import { isDataAccessorAstNode } from '@app-builder/models/astNode/data-accessor';
 
 import { getDataAccessorAstNodeField } from '../ast-node/getDataAccessorAstNodeField';
